+++ src/HttpBatchHandler/BatchMiddlewareExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using HttpBatchHandler.Events;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
