+++ src/HttpBatchHandler/Events/BatchMiddlewareEvents.cs
@@ -1,18 +1,8 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Http.Features;
 
-namespace HttpBatchHandler
+namespace HttpBatchHandler.Events
 {
-    public class BatchMiddlewareOptions
-    {
-        public PathString Match { get; set; } = "/api/batch";
-        public BatchMiddlewareEvents Events { get; set; } = new BatchMiddlewareEvents();
-    }
-
     public class BatchMiddlewareEvents
     {
         /// <summary>
@@ -54,35 +44,5 @@ namespace HttpBatchHandler
         /// After all batch request are executed
         /// </summary>
         public virtual Task BatchEnd(BatchEndContext context) => OnBatchEnd(context);
-    }
-
-    public class BatchEndContext
-    {
-        public MultipartWriter Writer { get; set; }
-        public int StatusCode { get; set; } = StatusCodes.Status200OK;
-        public bool WriteBody { get; set; } = true;
-        public Exception Exception { get; set; }
-        public object State { get; set; }
-    }
-
-    public class BatchRequestExecutingContext
-    {
-        public HttpApplicationRequestSection Request { get; set; }
-        public FeatureCollection Features { get; set; }
-        public object State { get; set; }
-    }
-
-    public class BatchStartContext
-    {
-        public object State { get; set; }
-    }
-
-    public class BatchRequestExecutedContext
-    {
-        public HttpApplicationRequestSection Request { get; set; }
-        public HttpApplicationMultipart Response { get; set; }
-        public bool Abort { get; set; }
-        public Exception Exception { get; set; }
-        public object State { get; set; }
     }
 }
