+++ src/HttpBatchHandler/Events/BatchStartContext.cs
@@ -1,0 +1,7 @@
+﻿namespace HttpBatchHandler.Events
+{
+    public class BatchStartContext
+    {
+        public object State { get; set; }
+    }
+}
